Fix lintings (#178)
* fix linting changes

* fixing other errors, like non-explict type

+++ contracts/L1GlobalForkRequester.sol
@@ -9,11 +9,8 @@ pragma solidity ^0.8.20;
 */
 
 import {IForkingManager} from "./interfaces/IForkingManager.sol";
-import {IPolygonZkEVMBridge} from "@RealityETH/zkevm-contracts/contracts/interfaces/IPolygonZkEVMBridge.sol";
-import {IPolygonZkEVM} from "@RealityETH/zkevm-contracts/contracts/interfaces/IPolygonZkEVM.sol";
 import {IForkonomicToken} from "./interfaces/IForkonomicToken.sol";
 import {IForkableBridge} from "./interfaces/IForkableBridge.sol";
-import {IBridgeMessageReceiver} from "@RealityETH/zkevm-contracts/contracts/interfaces/IBridgeMessageReceiver.sol";
 
 import {MoneyBox} from "./mixin/MoneyBox.sol";
 import {CalculateMoneyBoxAddress} from "./lib/CalculateMoneyBoxAddress.sol";
@@ -156,7 +153,7 @@ contract L1GlobalForkRequester {
         );
 
         if (doYesToken) {
-            uint newAmountToMigrate = amount -
+            uint256 newAmountToMigrate = amount -
                 failedRequests[token][requester][requestId].amountMigratedYes;
             failedRequests[yesToken][requester][requestId]
                 .amount += newAmountToMigrate;
